Remove function in Storage to display warning with Ui
The chatbot's I/O handling has changed from CLI to GUI. This means that Storage objects' interaction with Ui will no longer be displayed on the GUI.

All references in Storage to Ui are removed.

This is done in preparation to prevent future conflicts if and when Storage needs to propagate message towards the GUI instead.