Replace error logging messages used with InvalidCommand and KarenException to enum
Magic strings used to be scattered throughout the codebase. This makes it hard to track the error messages logged to user and hard to configure.

Hence, InvalidMessage acts as a config to store all error messages logged to user.

With the change, it is easier to track error logging messages in a single enum and allows for standardisation of messages.