Storage class taskToString & stringToTask methods handles tasks case-by-case on its own.
The methods are too long and unwieldy which can cause difficulty in
finding bugs later on.

Extracting out the methods based on the task type allows us to clearly
see what is being done based on the task's type.

This is to make the code look cleaner and make bug-finding easier.