Update start point for project
Project now is starting from the Launcher class.

This class is redundant as it is possible to move the application start point to another class in the existing project.

Let's remove the Launcher class, and move the main method into Main class.

Using this will enable the project to be more streamlined, and the naming conventions more readable.