Create assertions at critical parts
Project currently does not have assertion points at various parts of the code. Gradle does not have assertions enabled.

At runtime, it is difficult to assert which part of the codes are causing
run time issue.

Let's create assertion points at various parts of the code. We will create assertion points at
* Storage - when data is unable to load up successfully.
* Main - when window is not able to load successfully.
* TaskList - when the new object created is not created successfully.
This will allow us to check critical portions of the code and understand which
part of the codes are causing runtime error.