Refactor code to enhance code quality
Currently, some of the codes have deep nesting issue, and some of the methods are too long.

Need to resolve the deep nesting issue effectively in order to ensure better readability, and also to make the code easier to debug. Methods that are too long and have duplicate codes are shortened, some common functionalities are extracted out to stand alone methods.

Other than code refactoring, the following are done in this commit:
- Removed redundant methods that are not required anymore, i.e. isExit() method in command package
- Added enum along with switch case to resolve deep nesting issue.
- Added missing JavaDoc