Implement A-Assertions
Add assertions via the assert feature to Ui.java and Command classes
(AddToDoCommand.java, AddDeadlineCommand.java, AddEventCommand.java,
DeleteCommand.java, MarkCommand.java, UnmarkCommand.java,
PrintCommand.java, FindCommand.java).

Assertions allows program to terminate in the case of a possible bug,
and the sooner the execution stops, the safer it is.

Multiple checks for null objects or empty strings are done for the
parameters passed into the methods of Ui.java and the string response
returned from command executions.

Checking for null objects or empty strings will allow us to verify and
ensure that the program is behaving as expected.