Refactor code to improve code quality and readability
There are many benefits to improving code quality, which includes
easier maintenance of the codebase in the future in the event others
want to make changes to the program.

Some boolean checks were refactored to be self-explanatory. Comments
were added to some blocks of codes and methods to add clarity to the
code.

Some improvements in method length in the Parser class could be
good, but were not implemented in this commit yet as research is still
being done on the best way to refactor it.