Add welcome and goodbye message to GUI
The welcome and goodbye message of GUI is missing.

Users will not find this application interactive as it shows nothing at
the start and close abruptly at the end.

Adding a greeting message at the start makes the user feel more welcomed
and adding a goodbye message and a 1.5 second delay before closing will
make the experience of user more natural.