Move to IntelliJ, Create Jar, Fix data writing
Export Jar file, the jar file does not save the list on exiting.

Users downloading the released jar file need to save their progress
after exiting.

Let's,
* Fix the save bug and allow users to save their progress by changing
the output file to './data' instead of '/data'
* Export a JAR file for easy distribution
* Add more intuitive ways to mark a task as done such as 'do' and 'done'

The first fix is effective in removing the error of being unable to
save. Users usually refer to completing tasks as 'do' and 'done' instead
of 'mark'.