feat: add priority
There is no way to set priority levels for tasks.

Being able to set a priority level for each task will allow users to
note down which tasks are more urgent and therefore should be done first.

Let's add the ability to set a priority level for each task by
integrating into the Task class an Priority enum with three values:
* LOW
* MEDIUM
* HIGH

Using an enum is suitable because priority levels can be defined as a
set of predefined values, and using it over, say, a range of integer
values will also make the code easier to understand.

Refer to this website for more information on enum:
https://docs.oracle.com/javase/tutorial/java/javaOO/enum.html