A-Assertions: Use Assertions
Assertions not included

Assertions are used to define assumptions about the program state so
that the runtime can verify them. By adding assertions, the developer
can identify if the result is as intended, and can detect possible bugs
in the code in the event of an assertion failure.

Let's add insertions to check the intended behaviour of the task list
after adding, deleting, marking and unmarking commands.

Ensure that the task list will be properly modified upon the execution
of commands.