Add Java assertions
Currently, there are some assumptions made in the logic.

They may cause unintended consequences and regressions in the actual run.

Let's add several Java assertions to help us root the potential problems.

Using assertions is preferable because it narrows down the general error to a specific part of the code.