Add undo command for Duke
An undo command does not exist in the current duke.

Such a command is needed as users may want to undo an action that they made by mistake.

Add an undo command to allow users to undo add, mark, unmark and delete commands.

As all commands have different ways of undoing, it is better to override an undo method in the Command class and make use of polymorphism to customize each undo function.