Fix GUI to be functional for all commands
Current functionalities for GUI only allow a generic fixed response 'hi'
to any user input.

User would need the GUI to be able to respond with relevant messages
based on the commands input by the user.

Let's implement the run method of the Duke class to be able to return a
string of a specific format based on the command entered by the user. To
do this, we make the following main modifications:
    1. Create a command package to store all the different command
       classes.
    2. Create individual classes for each command type to implement the
       run() method to return a string of the required format - each of
       these command classes inherits from the generic Command class.
    3. Delete the Ui class which is no longer needed to interact with
       the user interface.