Updated GUI for better UI
GUI now has the ducky image. User image is also changed to an egg,
which fits with the yellow theme of the application.

Caught errors with the Parser class as it was not catching the case
where the user only mentions the command and nothing else. This was
causing the GUI to become buggy and resolving it was of top priority.

Assertions should be used alongside Exceptions in the classes to further
ensure that there will be no uncaught errors moving forward.