Add assertion to command types
Assertions help define assumptions about the commands added so that
runtime can verify them.

Asserts are added throughout the classes to ensure that the format of
user input is correct during runtime.

The sooner the execution stops, it is safer.

Refer to CS2103 website to find out more for the use of assertions
https://nus-cs2103-ay2122s2.github.io/website/se-book-adapted/chapters/errorHandling.html#assertions