Add base functionality for B-Reminders.
Currently Event and Deadline tasks have attached dates.

However, it is easy to forget these dates without reminders.

Let's add functionality to remind the user at a user-specified date
and time about the Events/Deadlines or even Todos if required.

A Reminder attribute has been added to Tasks to easily link Tasks to
Reminders. A method has also been added to the Parser to allow for
parsing Reminder commands.

This commit is the first part of multiple parts to implement the full
B-Reminders functionality. Users can now add Reminders to Tasks but the
functionality for actually reminding users once the reminder time is
reached has not been added.