Add some assert statements
There are no assertions in the code.

Some important assumptions about the code cannot be verified.

Let's use the assert feature to document important assumptions about
internal invariants as well as preconditions and postconditions.

Adding assert statements increases the robustness of the code because
any violation of assumptions about the code will be detected.