Add assertions
Modify build.gradle to allow assertions.

Assertion is added to storage as we need to ensure that the
txt file is created in order to store tasks in it.

Commands like mark, unmark, todo, delete and find are based
on single description (i.e integer/string), so we need to
make sure that it is non-empty description is passed.