Add assertions
After adding a task to the task list, program assumes that the size
of task list will always be more than 0.

Including assertions can help to define assumptions so that the
runtime can verify them.

Let's add an assertion to the add command so that we can verify that
the size of task list is always more than 0 after a task has been
added to it.