Add a places command
Additional functionality to include places in event task.

Have a choice to include a location into event. Bot can list out a list
of specific places if user intend to find all tasks at that location.

Having PlacesCommand so that command can vet if a task in list contains
the specific task.