Mark and Delete Command: throw InvalidIndexException
Mark and Delete Commands throw InvalidCommandException when the index portion of the command is left blank.

Both commands should throw the InvalidIndexException when index is not provided to give a clearer error message to the user.

Let's update the Parser.parseDelete and Parser.parseMark to throw the InvalidIndexException when applicable.