Duke.java: assertions used so define assumptions so that the runtime can verify them.
Currently, Duke is manually ensuring that all inputs have the right
number of tokens to ensure that the keywords used are provided the
appropriate amount of arguments.

This is inefficient and makes our code lengthier as a result.

We want to make use of java assert to verify these basic assumptions
rather then write specific if-else statements to handle this during
runtime.

Assert it a in-built function of Java that we can use to shorten our
code while still achieving the desired outcome.

As much as possible, all keywords are now asserting assumptions and
changes can be found within the run() method of Duke.java.