Add Tag and Untag features
Additional features such as tagging and untagging have been added
into Duke bot. This is so as to make the bot more versatile and
handle more user inputs.

These features give more flexibility to the bot and
users will also have additional functions to use when trying the
bot.

Implementing these features allow the user to be able to add
additional tags to their tasks to further express his/her
short comments or untag their comment. For example,
tagging/untagging can be done for all tasks that are present in
the user's list.

Let's create handleTag and handleUntag methods to process the logic
of these features in the Parser class.

Using composition is preferable over inheritance when adding the
Tag class into the Task class because a Task HAS A Tag. Thus,
Tag is an attribute of Task.