Added assertions
Added assertions to all Command classes and to the TaskList class. The
addition of assertion helps to ensure that the different tasks being
added are not empty. It also ensures that the description of the task is
not empty when passed into the Command classes for task creation. It
also helps to indicate a possible bug in the code as it violates an
assumption about how the code should be performed.

More assertions can be included in the Task classes to ensure that the
created classes have the necessary parameters to perform as expected.