Add assertion to code
We assumed the state of our variables especially when using private
functions.

Let's add assertion to our private function so that if a wrong argument
is passed into our private function, it will be flag out by assertion.

Using assertion instead of exceptions as private functions are not
invoked by others other than methods in the own class. These scenarios
are considered logically impossible, hence if the assertion is raise,
it is likely due to a mistake made by ourselves.