Add undo and redo functionality
Changes made to the list are permanent and have to be manually
undone or redone if the user wants to undo or redo them.

This makes changing the list tedious if the user makes a mistake.

Let's add undo and redo functionality by using two stacks to keep
track of the history of the tasks ArrayList in the history
taskList.

This is the simplest way to allow users to undo and redo any
changes made to the taskList.