Improve code quality
Code contains magic literals that is distracting the reviewer from the
actual implementation of the method.

Some codes also have the happy path nested, which is distracting as
those method are not the usual case.

Switch case default are being used to execute the last option, which
may cause unintended bugs.

Let's
* Ensure that there are no magic literals, replace all with constants
* Use guard clauses to ensure that happy path is prominent
* Remove unused methods and constants
* Use default branch, and to throw exception if there are no default
action

This would ensure that the code remains readable and easier to spot any
mistakes or bugs.