Refactor event and deadline class.
Various magic values in event and deadline class. Can cause confusion on what the values mean and make it harder to debug issues in the future.

Refactor them to no longer use magic numbers so it'll be clearer to understand and easier to change the values in the future. This will improve code quality.

Let's abstract out the magic values and declare them as named constants.