Remove complicated expressions.
Various long and complicated expressions in various command classes.

Long complicated expressions, makes it hard to read, understand and debug the code. It is also prone to bugs if the code were to get longer.

Let's abstract and seperat out the complicated expressions to simpler more readble expressions to improve code quality.