Add assertions to command class.
Command class getDescription function able to take in an empty input, or input without the proper command key.

Will cause a parsing bug and obtain the wrong data as the parsing of data requires heavily on the key being inside the input, or for the input to not be empty.

Let's add assertions to command class getDescription function, to ensure the key exists in the input passed in, and that the input is not empty.