Update test classes to most recent output
Test classes are not up to date.

This ensures future testing can be done when new features are added.

Let's edit the expected cases in the test classes to fit the new output.

Actual output contains a new line character at the end of a toString
method of a task. The change modifies expected to meet the new change
in the actual output.