Fix issues from code review
There are coding standard issues with previous commits.

Therefore, in order to fit with the standardised coding standard,
change the code to fit coding standard. Alter the switch cases within
Command.java as well as inappropriate method and variable names.

This is so all code will look as if they were written by
a single developer.