Add Statistics support
Implement stats feature to allow user to view their task statistics information.

Create a StatsCommand, UserProfile class to store the necessary information of the user and facilitate the displaying of the information in the UI.

Modify the storage system to store user information instead of only task list information.

By extracting the data conversion into their respective class of Task, TaskList, UserProfile, this allows organisation in the code.

Other info
- Update JavaDocs of StatsCommand and missing JavaDocs