Add the ability to detect task duplicates
Duplicated task (Deadline, ToDo, Event) can be added into the
TaskManager, which is redundant to the user.

This update prevents user from adding duplicated tasks.

Let's,
* add an abstract equals(Object) method to Task
* implement equals(Object) method in Task's subclasses
* add a duplicate check in addTask(TaskType, String) in TaskManager
* add generateDuplicateTaskMessage(TaskType) in Message

This uses the power of polymorphism by leveraging the use of the equal
method of Object and overriding it in subclasses, so minimal edits were
made.

Event objects are equal when description and at variables are equal.

Deadline objects are equal when description and at variables are equal.

ToDo objects are equal when description variables are equal