Improve code quality and OOP
Improve readability for reviewers as long methods, insufficient SLAP, and arrowhead style code is observed.

Added FileParser and InputParser class (subclass of Parser) to reduce the line of codes in the getResponse() method in Duke and loadFile() method in Storage. Renamed method names
to make it more intuitive for the code reviewers. Shifted methods (mainly the commands) from TaskList to Ui.

Unable to make Parser class universal without inheritance, which is why FileParser and InputParser subclass are created. Methods (mainly the commands) should not be placed in TaskList, but instead should be placed in Ui since this commands deals with input from the users and returns an output string to the interface.