Add gradlew and modify gradle.
Github Actions currently not working when it comes to building due to missing gradlew file. Current gradle file mainclassname is also not updated to the latest.

Need to add back the gradlew file and modify the gradle file so github action workflow will be able to run and this will make the work process much easier.

Let's add back the gradlew file.