Improved code quality
Removed magic numbers and strings to improve readability of the code.
Replaced certain if else conditions with booleans and named them based
on what they are supposed to show. e.g. isEmptyList.

Removed run method in Duke as it was no longer being used. More can be
done to improve the code quality. Some files have not been vetted for
magic numbers and strings