Reminder implementation with Timer only, no Thread.
In this implementation of the Reminder feature, the previous
implementation involving the use of Threads was scrapped as the
occurrence of the Thread in line with its sleep rate caused problematic
behavior.

Hence, in this implementation, the start & stop of the Reminder feature
is controlly only via Timer, with the schedule() & cancel() methods.