Refactor Task, Deadline, Event & TaskList.
We refactor Task, Deadline & Event so that Task has a LocalDate as one
of its variables so that calling getDate() can be applied to Task.

TaskList is also refactored so that the ArrayList containing the tasks
is a static object. After all, there will only be one instance of the
TaskList in the entire runtime of the application.