Add archive functionality
There is no way to tidy Duke's list unless the user deletes tasks individually. These deleted tasks are not stored.

Let's allow users to archive completed tasks so that they can:
- retain a tidy list of uncompleted tasks in a single command
- do not need to delete completed tasks 1 by 1
- can store completed tasks in a text file when archived

Only completed tasks are removed and sent to the archive as incomplete tasks should remain on the to-do list unless specifically deleted.