Refactor Codes
The codebase lacks documentation and has a lot of nested loops and
repeated parts.

Documentation helps people using the package have a clear understanding
of the methods and classes and use them correctly. It reduces future
development time. Replacing duplicate code makes the code more succinct
and easier to change implementation and debug in the future.