Add Update functionality
Tasks are updated by deleting existing tasks and repopulating
with an updated version of the same task.

This behaviour causes potential loss in information.

Creating an update command allows us to maintain the task objects
in the same list while changing the attributes in one command.

Let's add the update command to allow mutability of task objects.

This behaviour is prefereable because deleting objects and adding
copies of the same object is cumbersome and error prone.