Update the GUI
Label is used as the control to display text.

The use of label limits each message to 7 lines, and will become ...
after the message exceeds 7 lines (for example when listing more than
6 tasks).

Text is used instead to display the messages.

Text will dynamically resize the height when the message gets longer.

In addition, the photos of Puke and the user are updated.