showCommands(): make output more readable
showCommands()'s displayed message is unsightly and not user-friendly.

Edit showCommands() into for loop and change COMMANDS from type String
to String[]

Extracting COMMANDS into an array helps the showCommands() function to
print the list of commands to be updated when a new command is added
into the array, by using a for-loop.