Implement Increment C-DetectDuplicate
Users can add multiple tasks with the same name.

Users may accidentally add duplicate tasks without noticing, which
can cause confusion.

Let's detect when a duplicate task is added and give a warning
to the user.

Checking for duplicates is performed before the task is added so
the whole task list does not need to be iterated over to find
duplicates.