Update conventions for factory methods of Deadline, Todo and Event
Factory methods of Deadline, Event and Todo contains different extract-able logic and is very long.

The length of the code exceeds the recommended 30 lines and makes it harder to read.

Extracting the logic for recognizing the different attributes of the respective task from the user input into separate functions. 

Encapsulating the logic for recognizing separate attributes of a task makes the code more concise and readable.