Update conventions for Duke
Function initialize in Duke class have unclear paths between the erroneous/unusual and happy clauses of the function.

This violates the happy path prominent technique of java conventions and makes the code slightly unreadable as it can distract readers from the alternative paths.

Separated the logic for rare/error clauses (e.g. when a user uses the program initially) from the happy clause.

This makes the logic for the code more defined as it is properly split into separate sections. As a result the code is easier to read and understand.