Update conventions for Storage
Catch clause of function save does not accurately document it's intended behavior as it is common for this exception to be caught (e.g. when a user starts up for the first time).

Printing stack trace can be misinterpreted as a bug occurring in the program.

Removing the stack trace and leaving a documentation indicating that there is no problem would remove any ambiguity.

Because it is a commonly caught exception that does not raise any problem in the system, leaving a javadoc indicating so is sufficient and clear,