Add assert statements
User command information is assumed to be of the correct length after
parsing the input.

This might lead to Runtime Errors if user input is formatted wrongly.

Added assert statements after String splits to check if the given input
provides valid command information.

It is done this way to help us get familiarized with assert statements.