Add assertions for initialization checks
Added assertions for execution method in certain Task sub classes and Duke.

Task sub classes: Most of the execute functions in tasks requires and assumes that the task list provided is properly initialized. 

Duke functions: Functions in duke requires the internal managers of the duke to be initialized.

These assertions ensures that the functions are only called in the program loop.