Updated Duke class with new method greetUser()
greetUser() method functionality added to Duke class
allows for abstraction when the bot is started and a
welcome message needs to be displayed. This makes the
MainWindow class SLAP hard.

Let's add a greetUser() method which returns a string
greeting.

Doing so will make increase readability of the code
and makes debugging easier in the future. Also ensures
that relevant information is abstracted away.