Implement duplicated task feature
Duplicated task may cause confusion to users as it is difficult to
differentiate tasks of the same name.

Ensures that users do not add duplicated task with the same name
to the task list. In the event that when a duplicated task is added, a
warning is given instead. Informing the user that there is already a task
with the same name in the task list.

This feature aims to improve user experience by reducing user error.