Refactor code to improve the code quality
createCommand method in Parser is too long, and contains many repeated
lines of code. This is the same for the todayTask method in
DukeList.

Extracting the code logic allows us to use abstraction to  reduce
the amount of repeated code, and also increases the code readability.

Let's write new methods that contain the same code logic for different
commands in DukeParser.

Abstraction is preferable as code logic for certain commands with same
checks can be easily changed.