Add task priority feature
Task objects don't contain a priority attribute so there was no way to
assign priority to a task.

Some tasks might be more important and warrant more attention from the
user or even the program.

Added the priority variable in the Task class and updated the other
classes and methods to accommodate this change.

It is done this way because the best and most structured way to
implement this change is to use the existing OOP structure.