All create task commands: Create detection of duplicated dates
There are currently no checks on the task list to ensure that is a date
for a task before a new task is created.

To simulate real life events, two unrelated events cannot happen at the
same time. Hence it is important that we have this mechanism to check
that an event with the same date has been created before creation.

Let's create the algorithm on Deadline and Event to check whether there
is a deadline/event with the same date before task creation.