Include JavaFX in jar
In the event the JavaFX is not installed in the user's computer,
this will ensure that the necessary dependencies are packaged
together in the jar file. This allows the user to start using the
program without the need for additional downloads.

Change the main class to the launcher class in jar

Due to the implementation of JavaFX, this will ensure that the main
class is taken as the JavaFX launcher class. This will allow the
JavaFX GUI to be utilized.