Add Postpone feature
In previous versions, editing the dates for Event and Postpone tasks is not possible. The only way to do it is by deleting the task and adding it back with the correct date.

With this new version, the "postpone" keyword can be used to edit the date for those types of tasks. The command syntax is "postpone TASKNO DATE"