Update Storage and TaskList to include Assertions
Assertions are not used to check if tasks has indeed been added or deleted. Assertions are also not used to check if the directories has been created for the text file.

Using assertions asserts that tasks has been added for the various methods that add tasks to the tasklist. Also, it ensures that a new directory has indeed been created if the text file for storage does not exist.

Let's
Use Assertion to ensure that tasks have been added and deleted accordingly.
Use Assertions to ensure that the file directory has indeed been created.