Add assertions to ensure code integrity
Currently, the code does not make use of assertions to ensure
that assumptions hold in important methods. This makes the
program more prone to bugs.

Let's add assertions to methods that rely on assumptions
about arguments and about the contents of certain variables.

Assertions are a good choice for checking programmer assumptions.