Add assertion to functinal code in Storage and TaskList classes
In TaskList class, assertions had been made to the following methods:

  - addItem
  - deleteTask
  - markItem
  - unmarkItem

In Storage class, assertion had been made to the follow method:

  - load

These assertions are made to ensure that the methods performs the required
function that it is expected to do.

In addition, minor modifications had been made in the build.gradle file to
ensure that the jar file built is functional. These changes include modifying
the mainClassName of application to "siri.Launcher" as well as archiveBaseName
to "siri".