Added assertions throughout the code base.
Important assumptions mainly regarding validity of parameters exist in
various methods of the different classes in Duke.

Certain assumptions may not hold and that can cause the program to fail.

Let's add assertions at every part of the code where important
assumptions are being held to ensure that they hold true.

Using assertions to document assumptions within the program allows the
runtime to verify them. Upon assertion failure, the program typically
terminates with an error message and this prevents possible bugs in the
program from harming the program.