Refactor command classes to improve code quality
Currently, execute method in many command classes does not adhere
to SLAP, and does not make use of guard clauses.

This decreases code readability and maintainability.

Let's refactor the code so that all dialog-producing code
is handled in the Messages class instead, and guard clauses
are used when possible.

This allows for a clear separation of responsibilities amongst
classes and better adherence to SLAP.