Modify code to increase code quality
Modified the following classes:
  - Parser
  - Storage
  - TaskList

In Parser class, handleCommand function had been modified to increase code
quality as it had been previously lump with a single large switch case
making code readbility low. The following private functions had been
added to extract out the individual commands to be processed separately:

  - handleMarkCommand
  - handleUnmarkCommand
  - handleListCommand
  - handleTodoCommand
  - handleDeadlineCommand
  - handleEventComamand
  - handleDeleteCommand
  - handleEPrintCommand
  - handleDlPrintCommand
  - handleFindCommand

In TaskList class, JavaDocs had been modified for the following classes to
reflect the return value of the methods:

  - addItem
  - deleteItem
  - print
  - markItem
  - unmarkItem
  - printEventOn
  - printDeadlineOn
  - find

Additionally, changes had been made to addItem, deleteItem, markItem and
unmarkItem functions to process and throw list related exceptions such as
exception relating to list length as well as item index of list. A new
private function had been added to create a Deadline object base on the
loaded data so that other functions relying to create a new Deadline
object do not need to check whether time field is present. The error in
code for find had been fixed to from printing on console to returning
a string when tmp size is 0.

In Storage, load function had been modified to reduce the number of
indentation and make the code clearer.