.gitignore, Storage class: fix bug to track files in the data folder
.gitignore includes the path data/ which is supposed to refer to the local storage folder only, but is also the name of the package containing data classes such as Task.

The common name between the local storage folder and data package causes the data package to be ignored by git as the data/ path is included in .gitignore.

Changing the name of the local storage folder and replacing the data/ path in .gitignore with the new path to the local storage folder allows git to ignore the local storage folder only, while the data package is included in the git repository.

Let's rename the local storage folder to storageFolder, replace data/ in .gitignore with storageFolder/ and assign "storageFolder" to the constant DEFAULT_STORAGE_FOLDER_NAME in the Storage class.