Add assert statements to multiple project files
Possible bugs in the current code may lead to unexpected errors if not caught at the root.

Adding assert statements to various places in the code (e.g. in switch statements without a default case) allows the runtime to detect an assertion failure and terminate the execution with an error message.

Let's add assertion statements to places in the code which work only when certain assumptions about the program state are met.

Assertions are used instead of exceptions to check the internal logic of the code and ensure that it is correct, whereas exceptions are thrown for situations outside of the programmer's control.