Created new classes: ChangeMark to contain both Mark and Unmark.
MarkCommand and UnmarkCommand are very similar in nature, sharing many
lines of duplicated codes.

Duplicated codes contributes to a longer code and possibly more issues
arising due to more lines of code. Review process will be increasingly
tedious as well.

Let's abstract the functionality of Mark and Unmark to ChangeMark in
TaskList, Storage and Ui.