MainWindow.fxml: fix file name
Initially there was a typo in the name of the file, which causes the
Main class unable to load the file and throw errors in the jar file.

Change the name from "MainWIndow.fxml" to "MainWindow.fxml".

Otherwise, getResource() is not able to figure out the file and cannot
load it properly.