Add assertions: assert correct filepath
The operations on duke.storage.Storage.java relies on the correct file
path such that it can read/write on the correct data file. Similarly,
duke.main.Main.java relies on the correct fxml file for it to load the
GUI correctly.

Let's add assertions to these two classes. In duke.main.Main.java, add
an assertion to the start() method, such that it ensures that the
filePath is correctly loaded. In duke.storage.storage.readTasks, add an
assertion to ensure that the data file is correctly loaded.

Using assertions ensure that the GUI can be loaded correctly and user
will not loss his/her data.