Completed B-DetectAnomalies
Add new extensions

Added the following extensions to the project:

  - Detect anomalies when task is being added (clashes between tasks)
  - Mass operations allowed for delete, mark and unmark

In Event class, new getters had been set to get the date and time for the
event item.

In Parser class, a private function had been create to generate a list of
interger from a string passed.

In TaskList class, deleteItem, markItem and unmarkItem had been modified to allow
manipulation of multiple task at one time. In addItem, error will be thrown if
user want to add a new event and the date and time of event clashes with existing
event.