Add sort functionalities for task list
Requirement to be able to sort the task list according to alphabetical and chronological order.

Functionality allows user to sort and view their task in order.

Add two different sorting functions, one for chronological and one for alphabetical.

This is the two most common ways users would want to view their tasks by.