Added tagging with storage
Task class: Stores task data such as name/description, timing data,
mark/unmark status.

Tasks can be categorised into different priority levels (e.g. urgent,
important, optional, etc).

There are no possible way to categorise tasks.

Adding a list of tags to the Task class allows tasks to be categorised
into various different categories. For example, an event can be tagged
as "optional" and "expiring" and thus fit in both categories.

Let's add an arraylist of strings to represent the tags associated with
the Task class.

Using an arraylist of strings to store the tags is a convenient way
of storing and displaying tags of tasks. May create a dedicated Tag
class in the future instead.