Add Assertions and Exceptions
The current behavior is that Tsohg will not provide any response when index is out of range for certain commands.

This has to be changed for better user experience.

Exception and more error messages have been added to handle the scenario.

Assertions are also added to ensure the program works well.