Refactor Command parsing to Parser class
Parsing logic for each type of command is handled by the respective
command classes.

This makes it difficult to create instances of the commands.

Extracting the parsing logic of these commands into Parser class
makes the parser class the only class responsible for understanding
user input.

Let's:
* Remove all logic dealing with user input in each command class and
put them in parser class.
* Create the fields needed by each command as private variables.

Dead code and classes are also removed in this commit.