Improve code quality: extract various exceptions
Exceptions in the Parser class have similar structure and behaviour.

Using a command class IncorrectCommand is inaccurate, as errors
are not executed and hence should not be handled using a command class.

Let's pull out errors in the parser that have behaviour in common into
new exceptions.

Using inheritance for the subclasses of DukeException over composition
is preferred as the exceptions have different error messages.