Include assertions: add few assertions in the program
There are several assumptions made while running the program.

There are might come situations when these assumptions are not met, but
the program would still be running.

We need a way to alert, the programmer that the program is being run as
expected. Let's add assertion statements at various points in the code,
to make sure our assumptions are met.

This way we would know where we went wrong and what exactly was missing.